// src/views/Show/types.ts
export interface IGenre {
    id: number;
    name: string;
}

export interface IMovieResponse {
    adult: boolean;
    backdrop_path: string;
    genres: IGenre[]; 
    genre_ids: number[];
    id: number;
    original_language: string;
    original_title: string;
    overview: string;
    popularity: number;
    poster_path: string;
    release_date: Date;
    title: string;
    video: boolean;
    vote_average: number;
    vote_count: number;
}